+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -295,7 +295,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-03-17T08:28:29.867Z",
+    "2025-03-17T09:50:25.442Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -335,7 +335,7 @@ export const ListDatasetsData$outboundSchema: z.ZodType<
   updatedById: z.string().optional(),
   metadata: z.lazy(() => ListDatasetsMetadata$outboundSchema),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-03-17T08:28:29.867Z"))
+  updated: z.date().default(() => new Date("2025-03-17T09:50:25.442Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {