+++ packages/orq-rc/src/sdk/files.ts
@@ -15,7 +15,7 @@ export class Files extends ClientSDK {
    * Create file
    *
    * @remarks
-   * Files are used to upload documents that can be used with features like [Deployments](https://docs.orq.ai/reference/post_v2-deployments-get-config).
+   * Files are used to upload documents that can be used with features like [Deployments](https://docs.orq.ai/reference/deploymentinvoke-1).
    */
   async create(
     request?: operations.FileUploadRequestBody | undefined,