+++ packages/orq-rc/RELEASES.md
@@ -909,3 +909,13 @@ Based on:
 - [typescript v3.2.0-rc.51] packages/orq-rc
 ### Releases
 - [NPM v3.2.0-rc.51] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.51 - packages/orq-rc
+
+## 2025-03-17 09:51:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.519.0 (2.552.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.2.0-rc.52] packages/orq-rc
+### Releases
+- [NPM v3.2.0-rc.52] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.52 - packages/orq-rc